requirements.txt
discord.py==2.5.0
python-dotenv==1.0.1
requests==2.32.3
yfinance==0.2.54
pandas>=2.0.0